fix: reset TabNavigator state when going back from cross tab navigation
Ref #333
